+++ lib/pymedphys/_experimental/monomanage/app/api.py
@@ -26,7 +26,6 @@
 
 import os
 
-
 from .wheels import build_wheels_with_yarn, copy_wheels
