+++ lib/pymedphys/_experimental/monomanage/draw/directories.py
@@ -25,19 +25,18 @@
 
 
 import os
+from copy import copy
 
 import networkx as nx
-from copy import copy
 
 from ..tree import PackageTree
 from .utilities import (
-    save_dot_file,
-    remove_prefix,
-    get_levels,
-    create_labels,
     create_href,
+    create_labels,
+    get_levels,
+    remove_prefix,
+    save_dot_file,
 )
-
 
 ROOT = os.getcwd()
