+++ lib/pymedphys/_experimental/monomanage/draw/files.py
@@ -26,22 +26,20 @@
 
 import os
 import textwrap
+from copy import copy
 
 import networkx as nx
-from copy import copy
 
 from ..tree import PackageTree
 from .utilities import (
-    save_dot_file,
-    remove_prefix,
-    get_levels,
-    remove_prefix,
-    remove_postfix,
     convert_path_to_package,
-    create_labels,
     create_href,
+    create_labels,
+    get_levels,
+    remove_postfix,
+    remove_prefix,
+    save_dot_file,
 )
-
 
 ROOT = os.getcwd()
