+++ lib/pymedphys/_experimental/monomanage/parse/imports.py
@@ -23,8 +23,8 @@
 # You should have received a copy of the Apache-2.0 along with this
 # program. If not, see <http://www.apache.org/licenses/LICENSE-2.0>.
 
-import os
 import ast
+import os
 
 from stdlib_list import stdlib_list
