+++ lib/pymedphys/_experimental/monomanage/tree/__init__.py
@@ -1,2 +1,2 @@
-from .build import build_tree, test_tree, PackageTree
+from .build import PackageTree, build_tree, test_tree
 from .check import is_imports_json_up_to_date, update_imports_json
